pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "viky", email = "vikyw89@gmail.com"},
]
dependencies = [
    "easycodefpy>=1.0.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

